+++ README.md
@@ -98,15 +98,15 @@ npm run dev
       - 하루 지하철 이용자 수 400만명으로 잡음
       - 출퇴근 해서 400 * 2 = 800만
       - 8000000 / 86,400 (초/일) = 92.6 rps(1일 평균rps)
-      - 피크 시간대 집중률 = 3
-      - 1일 최대 rps = 92.6 * 3 = 277.8 rps
+      - 피크 시간대 집중률 = 10
+      - 1일 최대 rps = 92.6 * 10 = 926 rps
     - 부하 유지기간 : 30분
   - 부하 테스트 시 저장될 데이터 건수 및 크기
-    - http_req_duration = 8ms
-    - T(VU iteration) = (2 * 0.18) = 0.36s
+    - http_req_duration = 200ms
+    - T(VU iteration) = (2 * 0.2) = 0.4s
     - 목표 VUSER
-      - 평균 (92.6 * 0.36) / 2 = 16.7
-      - 최대 16.7 * 3 = 50.1
+      - 평균 (92.6 * 0.4) / 2 = 18.52
+      - 최대 18.52 * 10 = 185.2
 - 각 시나리오에 맞춰 스크립트 작성
   - 접속 빈도가 높은 페이지
     - 즐겨찾기 페이지
@@ -116,13 +116,27 @@ npm run dev
     - path 페이지- lineService, stationService, pathService 3개의 서비스를 참조한다.
 2. Smoke, Load, Stress 테스트 스크립트와 결과를 공유해주세요
 - 테스트 스크립트는 loadtest폴더 안에 있음.
-- Smoke 테스트 결과
-![smoke_result.png](loadtest/smoke_result.png)
-- Load 테스트 결과
-![load_result.png](loadtest/load_result.png)
-- Stress 테스트 결과
-![stress_result.png](loadtest/stress_result.png)
-
+- 접속 빈도가 높은 페이지(즐겨찾기 페이지)
+  - Smoke 테스트 결과
+  ![smoke_result.png](loadtest/getfavorites/smoke_result.png)
+  - Load 테스트 결과
+  ![load_result.png](loadtest/getfavorites/load_result.png)
+  - Stress 테스트 결과
+  ![stress_result.png](loadtest/getfavorites/stress_result.png)
+- 데이터를 갱신하는 페이지(내 정보 수정 페이지)
+  - Smoke 테스트 결과
+    ![smoke_result.png](loadtest/editmyinfo/smoke_result.png)
+  - Load 테스트 결과
+    ![load_result.png](loadtest/editmyinfo/load_result.png)
+  - Stress 테스트 결과
+    ![stress_result.png](loadtest/editmyinfo/stress_result.png)
+- 데이터를 조회하는데 여러 데이터를 참조하는 페이지(경로탐색 페이지)
+  - Smoke 테스트 결과
+    ![smoke_result.png](loadtest/findpath/smoke_result.png)
+  - Load 테스트 결과
+    ![load_result.png](loadtest/findpath/load_result.png)
+  - Stress 테스트 결과
+    ![stress_result.png](loadtest/findpath/stress_result.png)
 ---
 
 ### 3단계 - 로깅, 모니터링
